+++ inst/examples/example_layout_rename_ph_labels.R
@@ -16,12 +16,16 @@ layout_rename_ph_labels(x, "Comparison", "Title 1" = "LABEL MATCHED") # label ma
 layout_rename_ph_labels(x, "Comparison", "3" = "ID MATCHED") # id matching
 layout_rename_ph_labels(x, "Comparison", "Date Placeholder 6" = "DATE", "8" = "FOOTER") # label and id
 
+# rename using a named list and the .dots arg
+renames <- list("Content Placeholder 3" = "CONTENT_1", "6" = "CONTENT_2")
+layout_rename_ph_labels(x, "Comparison", .dots = renames)
+
 # rename via rhs assignment and optional index (not id!)
 layout_rename_ph_labels(x, "Comparison") <- LETTERS[1:8]
 layout_rename_ph_labels(x, "Comparison")[1:3] <- paste("CHANGED", 1:3)
 
 # rename via rhs assignment and ph id (not index)
-layout_rename_ph_labels(x, "Comparison", id = c(2,4)) <- paste("ID =", c(2,4))
+layout_rename_ph_labels(x, "Comparison", id = c(2, 4)) <- paste("ID =", c(2, 4))
 
 
 # MORE ------------
@@ -44,4 +48,3 @@ layout_rename_ph_labels(x, "2-dupes")[idx] <- paste("DUPLICATE", seq_along(idx))
 # warning: in case of duped labels only the first occurrence is renamed
 x <- read_pptx(file)
 layout_rename_ph_labels(x, "2-dupes", "Content 7" = "new label")
-